switch `consume_next` to `&mut self`, avoid unsafe `ptr::replace`
encountered a panic running examples/ws_client.rs in --release mode,
(but not in debug mode). when I changed this, it fixed the panics.
Possibly the `ptr::replace` usage was unsound (not sure, unsafe
semantics are pretty tricky). In any event seems like this is a pretty
reasonable change, seems like `consume_next` should have an `&mut self`
signature anyways, and changing the signature required no other changes
to build.